# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HireSync AI - Job Finder Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs/search:
    get:
      operationId: searchJobs
      tags: [jobs]
      summary: Search jobs
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: remote
          in: query
          schema:
            type: string
        - name: experience
          in: query
          schema:
            type: string
        - name: salary_min
          in: query
          schema:
            type: number
        - name: salary_max
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Job search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
  /jobs/recommended:
    get:
      operationId: getRecommendedJobs
      tags: [jobs]
      summary: Get AI-recommended jobs for user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recommended jobs with match scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendedJob"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    put:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All notifications marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Job:
      type: object
      properties:
        id:
          type: string
        externalId:
          type: string
        title:
          type: string
        company:
          type: string
        companyLogo:
          type: string
        location:
          type: string
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        salaryCurrency:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        source:
          type: string
        sourceUrl:
          type: string
        postedAt:
          type: string
          format: date-time
        employmentType:
          type: string
          enum: [fulltime, parttime, contract, internship, freelance]
        experienceLevel:
          type: string
          enum: [fresher, junior, mid, senior, lead, any]
        isActive:
          type: boolean
        isSaved:
          type: boolean
        hasApplied:
          type: boolean
      required:
        - id
        - title
        - company
        - location
        - description
        - source
        - sourceUrl
        - postedAt
        - isActive
    JobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - jobs
        - total
        - page
        - totalPages
    RecommendedJob:
      type: object
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            matchScore:
              type: integer
              minimum: 0
              maximum: 100
            matchedSkills:
              type: array
              items:
                type: string
            missingSkills:
              type: array
              items:
                type: string
            matchReasons:
              type: array
              items:
                type: string
          required:
            - matchScore
            - matchedSkills
            - missingSkills
            - matchReasons
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [job_match, application_update, alert, system, resume_score]
        isRead:
          type: boolean
        jobId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
...